Перенес SelectablePrintDoc из проекта с ГТК и удалил ненужный метод, использующий его
